Add mix format check job to elixir-build-checks workflow
[Card](https://rentpath.atlassian.net/browse/SRV-4854)

These changes add a new `format` job to the elixir-build-checks workflow
that validates the formatting of changed files matches what `mix format`
expects. We only validate the formatting of changed files because right
now our existing code deviates significantly from what the formatter
expects and we do not want to invest the time in updating the format of
all code at once.

Details on the `--no-run-if-empty` flag: I didn't catch this issue
initially because on Mac OS xargs does not run the provided command by
default when input is empty. On all versions of xargs installed on other
linux OSs the --no-run-if-empty command is required. See
https://stackoverflow.com/a/8296746 for details.