Pass GitHub PAT into actions-services-elixir-deps
[Card](https://rentpath.atlassian.net/browse/SRV-4904)

The actions-services-elixir-deps now accepts a github-token input that
needs to be set to the GHCR_USER_PAT secret value if we want to be able
to download dependencies from private ("internal") Github repositories.
Passing in the secret is necessary because actions cannot access secrets
directly.

The GHCR_USER_PAT is also passed in as a build argument for the Docker
build, because dependencies will need to be downloaded inside the
image build as well.